+++ examples/term_to_ic.rs
@@ -7,10 +7,11 @@ fn main() {
 
     for term in ontology.hpos() {
         println!(
-            "{}\t{}\t{}",
+            "{}\t{}\t{}\t{}",
             term.id(),
             term.information_content().gene(),
             term.information_content().omim_disease(),
+            term.information_content().orpha_disease(),
         );
     }
 }
@@ -23,15 +24,15 @@ _ = Ontology()
 
 with open("term2ic.py.txt", "w") as fh:
     for term in Ontology:
-        _ = fh.write(f"{term.id}\t{term.information_content.gene}\t{term.information_content.omim}\n")
+        _ = fh.write(f"{term.id}\t{term.information_content.gene}\t{term.information_content.omim}\t{term.information_content.orpha}\n")
 
 */
 
 /*
-Diff'ing (only use the first 3 decimal digits to ignore rounding errors:
+Diff'ing (only use the first 2 decimal digits to ignore rounding errors:
 
 diff \
-<(awk '{print $1, substr($2,1,5), substr($3,1,5)}' example_data/term2ic.rs.txt) \
-<(awk '{print $1, substr($2,1,5), substr($3,1,5)}' example_data/term2ic.py.txt)
+<(sort example_data/term2ic.rs.txt | awk '{print $1, substr($2,1,4), substr($3,1,4), substr($4,1,4)}') \
+<(sort example_data/term2ic.py.txt | awk '{print $1, substr($2,1,4), substr($3,1,4), substr($4,1,4)}')
 
 */