+++ src/similarity.rs
@@ -312,7 +312,7 @@ impl SimilarityCombiner for StandardCombiner {
 ///     StandardCombiner::default()
 /// );
 ///
-/// assert_eq!(similarity, 0.8177036);
+/// assert_eq!(similarity, 0.695935);
 /// ```
 ///
 /// ## Using `GroupSimilarity` directly
@@ -354,7 +354,7 @@ impl SimilarityCombiner for StandardCombiner {
 ///     Builtins::GraphIc(InformationContentKind::Omim)
 /// );
 ///
-/// assert_eq!(sim.calculate(&set_1, &set_2), 0.8177036);
+/// assert_eq!(sim.calculate(&set_1, &set_2), 0.695935);
 /// ```
 pub struct GroupSimilarity<T, C> {
     combiner: C,
@@ -456,9 +456,9 @@ impl Default for GroupSimilarity<GraphIc, StandardCombiner> {
 pub enum Builtins {
     /// [Distance](`Distance`) - based similarity
     Distance(InformationContentKind),
-    /// [GraphIc](`GraphIc`) - based similarity
+    /// [`GraphIc`](`GraphIc`) - based similarity
     GraphIc(InformationContentKind),
-    /// [InformationCoefficient](`InformationCoefficient`) - based similarity
+    /// [`InformationCoefficient`](`InformationCoefficient`) - based similarity
     InformationCoefficient(InformationContentKind),
     /// [Jiang & Conrath](`Jc`) - based similarity
     Jc(InformationContentKind),