+++ CHANGELOG.md
@@ -2,7 +2,16 @@
 
 All notable changes to this project will be documented in this file.
 
+## Unreleased
+
 ## [0.10.0]
+
+### Feature
+
+- Add Orphante diseases (`OrphaDisease`) to Ontology
+- Filter gene and disease annotations in subontology based on association with phenotypes
+- Add binary version 3
+- Add new example ontology
 
 ### Documentation
 
@@ -11,6 +20,9 @@ All notable changes to this project will be documented in this file.
 ### Refactor
 
 - Improve the OBO parser with better error handling
+- [**breaking**] Add `Disease` trait that is needed to work with `OmimDisease` and `OrphaDisease`
+- Update example ontology
+- Update unit- and doctests to align with updated example ontology
 
 
 ## [0.9.1] - 2024-03-30